+++ src/hyperlight_guest_bin/src/arch/aarch64/mod.rs
@@ -39,7 +39,7 @@ pub mod dispatch {
     );
 }
 
-mod exception;
+pub(crate) mod exception;
 
 macro_rules! msr {
     ($sysreg:ident, $expr:expr) => {