+++ src/hyperlight_guest_bin/src/lib.rs
@@ -23,10 +23,7 @@ use hyperlight_guest::guest_handle::handle::GuestHandle;
 #[cfg_attr(target_arch = "x86_64", path = "arch/amd64/mod.rs")]
 #[cfg_attr(target_arch = "aarch64", path = "arch/aarch64/mod.rs")]
 mod arch;
-// temporarily expose the architecture-specific exception interface;
-// this should be replaced with something a bit more abstract in the
-// near future.
-#[cfg(target_arch = "x86_64")]
+// Temporarily expose architecture-specific exception interfaces.
 pub mod exception;
 pub mod guest_function {
     pub(super) mod call;